psx_tracker/requirements.txt
psx-data-reader
schedule
pandas
beautifulsoup4
requests
